requirements.txt
absl-py==2.1.0
attr==0.3.1
attrs==24.2.0
coverage==7.5.3
lxml==5.3.0
pyserial==3.5
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.2

